Refactoring location of data source and using NewProtocol6WithError for tests (#93)
