+++ internal/provider/provider_test.go
@@ -9,7 +9,7 @@ import (
 func protoV6ProviderFactories() map[string]func() (tfprotov6.ProviderServer, error) {
 	return map[string]func() (tfprotov6.ProviderServer, error){
 		"http": func() (tfprotov6.ProviderServer, error) {
-			return providerserver.NewProtocol6(New())(), nil
+			return providerserver.NewProtocol6WithError(New())()
 		},
 	}
 }
